pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "highway-env>=1.10.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib",
    "gymnasium",
    "highway_env",
    "pytorch",
    "numpy"
]
